requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
filelock==3.11.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
orjson==3.8.9
protobuf==3.20.3
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
snscrape==0.6.2.20230320
soupsieve==2.4
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
